雷达监控。/quicktoolshub-python/backup_20251225_130645_Gold_V2_Homepage_Revamp/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.0.0
PyPDF2==3.0.1
pdf2docx==0.5.8
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
markdown==3.5.1
rembg==2.0.50
onnxruntime==1.16.3
werkzeug==3.0.1
